re-arrangement of files. Classes, Extensions, Functions, Structs and Views with single use files like Functions.swift broken out to seperate files in groups
added extra `--key` option. swiftDialog will read in a user preference from the `au.csiro.dialog` domain representing a sha256 hash of some secret phrase and compare against the `--key` value. If a value is present and `--key` is missing or does not match, dialog will refuse to launch.
added extra `--checksum` option that converts the supplied string to a sha256 hash value (provided for convenience. The same effect can be had by running `echo -n "My Secret" | shasum -a 256`)
