+++ dialog/Extensions/NSWindow+Additions.swift
