+++ dialog/Functions/CLOptions.swift
@@ -1,5 +1,5 @@
 //
-//  Options.swift
+//  CLOptions.swift
 //  dialog
 //
 //  Created by Bart Reardon on 10/3/21.
