+++ dialog/Structs/HelpText.swift
