+++ dialog/Structs/ImageProcessing.swift
