+++ dialog/Structs/jamfHelper.swift
