+++ dialog/Views/CheckboxView.swift
