+++ dialog/Views/ConstructionKit/CKWindowProperties.swift
