+++ dialog/Views/ErrorView.swift
