+++ dialog/Views/HelpView.swift
