+++ dialog/Views/ImageSlider.swift
