+++ dialog/Views/InfoBoxView.swift
