+++ dialog/Views/MessageContentView.swift
