+++ dialog/Views/MiniView.swift
