+++ dialog/Views/TaskProgressView.swift
