+++ dialog/Views/TextEntryView.swift
