+++ dialog/Views/WatermarkView.swift
